# 8339738: RISC-V: Vectorize crc32 intrinsic