… src/hotspot/cpu/riscv/stubRoutines_riscv.cpp