Hi,
Can you help to review this patch?
Thanks.

This improvement is based on java.base/share/native/libzip/zlib/zcrc32.c, I made some modification to N (to 16) related code, then re-generate the tables needed, finally vectorize the code (original implementation in zcrc32.c is just scalar code).

## Test
test/hotspot/jtreg/compiler/intrinsics/zip/TestCRC32.java, 
test/jdk/java/util/zip/TestCRC32.java

## Performance

### on bananapi

with patch
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark | (count) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | --
TestCRC32.testCRC32Update | 64 | avgt | 10 | 222.297 | 0.106 | ns/op
TestCRC32.testCRC32Update | 128 | avgt | 10 | 365.144 | 0.196 | ns/op
TestCRC32.testCRC32Update | 256 | avgt | 10 | 687.14 | 0.235 | ns/op
TestCRC32.testCRC32Update | 512 | avgt | 10 | 1043.833 | 0.083 | ns/op
TestCRC32.testCRC32Update | 2048 | avgt | 10 | 3299.928 | 1.361 | ns/op
TestCRC32.testCRC32Update | 16384 | avgt | 10 | 24384.502 | 25.298 | ns/op
TestCRC32.testCRC32Update | 65536 | avgt | 10 | 103200.458 | 8.297 | ns/op

</google-sheets-html-origin>

without patch
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark | (count) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | --
TestCRC32.testCRC32Update | 64 | avgt | 10 | 220.878 | 0.02 | ns/op
TestCRC32.testCRC32Update | 128 | avgt | 10 | 364.173 | 0.032 | ns/op
TestCRC32.testCRC32Update | 256 | avgt | 10 | 685.815 | 0.055 | ns/op
TestCRC32.testCRC32Update | 512 | avgt | 10 | 1329.049 | 0.084 | ns/op
TestCRC32.testCRC32Update | 2048 | avgt | 10 | 5189.302 | 0.666 | ns/op
TestCRC32.testCRC32Update | 16384 | avgt | 10 | 41250.873 | 23.882 | ns/op
TestCRC32.testCRC32Update | 65536 | avgt | 10 | 171664.002 | 15.011 | ns/op

</google-sheets-html-origin>

### on K230

with patch
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark with patch ( on k230) | (count) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | --
TestCRC32.testCRC32Update | 64 | avgt | 10 | 227.003 | 1.262 | ns/op
TestCRC32.testCRC32Update | 128 | avgt | 10 | 373.193 | 1.532 | ns/op
TestCRC32.testCRC32Update | 256 | avgt | 10 | 711.857 | 31.867 | ns/op
TestCRC32.testCRC32Update | 512 | avgt | 10 | 1363.103 | 8.453 | ns/op
TestCRC32.testCRC32Update | 2048 | avgt | 10 | 4601.248 | 18.298 | ns/op
TestCRC32.testCRC32Update | 16384 | avgt | 10 | 35536.58 | 764.026 | ns/op
TestCRC32.testCRC32Update | 65536 | avgt | 10 | 146814.093 | 522.202 | ns/op

</google-sheets-html-origin>

without patch
<google-sheets-html-origin style="caret-color: rgb(0, 0, 0); color: rgb(0, 0, 0); font-style: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: auto; text-align: start; text-indent: 0px; text-transform: none; white-space: normal; widows: auto; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration: none;">
Benchmark - without patch | (count) | Mode | Cnt | Score | Error | Units
-- | -- | -- | -- | -- | -- | --
TestCRC32.testCRC32Update | 64 | avgt | 10 | 225.475 | 1.434 | ns/op
TestCRC32.testCRC32Update | 128 | avgt | 10 | 371.716 | 2.301 | ns/op
TestCRC32.testCRC32Update | 256 | avgt | 10 | 700.833 | 3.693 | ns/op
TestCRC32.testCRC32Update | 512 | avgt | 10 | 1360.036 | 6.684 | ns/op
TestCRC32.testCRC32Update | 2048 | avgt | 10 | 5315.038 | 21.806 | ns/op
TestCRC32.testCRC32Update | 16384 | avgt | 10 | 42699.086 | 594.981 | ns/op
TestCRC32.testCRC32Update | 65536 | avgt | 10 | 177882.278 | 2511.731 | ns/op

</google-sheets-html-origin>

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8339738](https://bugs.openjdk.org/browse/JDK-8339738): RISC-V: Vectorize crc32 intrinsic (**Enhancement** - P4)


### Reviewers
 * [Fei Yang](https://openjdk.org/census#fyang) (@RealFYang - **Reviewer**)
 * [Ludovic Henry](https://openjdk.org/census#luhenry) (@luhenry - Committer)

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/20910/head:pull/20910` \
`$ git checkout pull/20910`

Update a local copy of the PR: \
`$ git checkout pull/20910` \
`$ git pull https://git.openjdk.org/jdk.git pull/20910/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 20910`

View PR using the GUI difftool: \
`$ git pr show -t 20910`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/20910.diff">https://git.openjdk.org/jdk/pull/20910.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/20910#issuecomment-2337749589)