@Hamlin-Li 2024-09-16
> Thanks for the update. Several minor comments remain. BTW: Do you have the updated JMH data to see?

Thanks for the detailed review.
With the latest patch, jmh shows no much different result.